Prow integration with Airflow is working.
* Airflow now successfully uploads test results (junit files) to GCS
  to support Gubernator.

* airflow.py now properly waits for the pipeline to terminate.
* Add some unittests.
* Renamed runner.py to prow.py; we no longer need runner.py since
  runner.py is replaced by our Airflow pipeline to run all our tests.
