+++ py/airflow_test.py
@@ -1,0 +1,34 @@
+"""Unittest for airflow."""
+import datetime
+import tempfile
+import unittest
+
+import mock
+import yaml
+
+from py import airflow
+
+
+class AirflowTest(unittest.TestCase):
+  @mock.patch('py.airflow.prow')
+  @mock.patch('py.airflow.wait_for_tf_k8s_tests')
+  @mock.patch('py.airflow.trigger_tf_k8s_tests_dag')
+  def test_main(self, mock_trigger, mock_wait, mock_prow):
+    mock_trigger.return_value = ("2017-11-01T12:12:12", "")
+    mock_wait.return_value = "success"
+    airflow.main()
+
+  def test_wait_for_tf_k8s_tests(self):
+    client = mock.MagicMock(spec=airflow.AirflowClient)
+    client.get_task_status.side_effect = [
+        { "state": None},
+        { "state": "running"},
+        { "state": "success"},
+      ]
+
+    state = airflow.wait_for_tf_k8s_tests(client, "2017-11-01T12:12:12",
+                                          polling_interval=datetime.timedelta(seconds=0))
+
+    self.assertEquals("success", state)
+if __name__ == "__main__":
+  unittest.main()
